% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knots_find}
\alias{knots_find}
\title{updated NR methods to find knots
This is a function about how to find knots in spline linear regression}
\usage{
knots_find(Y, X, tau_iter, epsilon = 1e-06)
}
\arguments{
\item{Y}{a vector being the observations of the response variable}

\item{X}{a matrix with the first column being the variable of interest and the left column being the covariates}

\item{tau_iter}{a vector being the estimated knots}

\item{epsilon}{threshold for error}
}
\description{
updated NR methods to find knots
This is a function about how to find knots in spline linear regression
}
\examples{
library(lsknots)
set.seed(12345)
n = 1000
eta = 0.5 
sig = 0.03
a=  rnorm(n,0,1)
Z = rnorm(n,0,2) # Z: confounder/other covariate
u = (a+Z)/sqrt(5)
X= pnorm(u,mean=0,sd=1) #X: factor of interest with change-point effects: uniform (0,1)
e = rnorm (n,0,sd=sig)  #e: error term
Y = theta[1] + theta[2]*X + theta[3] * as.numeric(X > theta[7]) * ( X - theta[7] ) + theta[4] * as.numeric(X > theta[8]) * ( X - theta[8] ) + theta[5] * as.numeric(X > theta[9]) * ( X - theta[9] )  + theta[6] * as.numeric(X > theta[10]) * ( X - theta[10] )+ eta*Z+ e 
tau00=c(0.3,0.5,0.85,0.95)
A=matrix(c(X,Z),ncol=2)
knots_find(Y,A,tau00)
}
