% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knots_ols}
\alias{knots_ols}
\title{OLS with estimated knots
This is a function about how to calculate the OLS estimator with given knots.}
\usage{
knots_ols(Y, X, tau_iter)
}
\arguments{
\item{Y}{a vector being the observations of the response variable}

\item{X}{a matrix with the first column being the variable of interest and the left column being the covariates}

\item{tau_iter}{a vector being the estimated knots}
}
\value{
estimate of the coefficients and response variable
}
\description{
OLS with estimated knots
This is a function about how to calculate the OLS estimator with given knots.
}
\examples{
library(lsknots)
set.seed(12345)
n = 1000
eta = 0.5 
sig = 0.03
a=  rnorm(n,0,1)
Z = rnorm(n,0,2) 
u = (a+Z)/sqrt(5)
X= pnorm(u,mean=0,sd=1) 
e = rnorm (n,0,sd=sig)  
Y = theta[1] + theta[2]*X + theta[3] * as.numeric(X > theta[7]) * ( X - theta[7] ) + theta[4] * as.numeric(X > theta[8]) * ( X - theta[8] ) + theta[5] * as.numeric(X > theta[9]) * ( X - theta[9] )  + theta[6] * as.numeric(X > theta[10]) * ( X - theta[10] )+ eta*Z+ e 
A=matrix(c(X,Z),ncol=2)
tau00=c(0.3,0.5,0.85,0.95)
yhat =knots_ols(Y,A,tau0)$y_hat
betahat=knots_ols(Y,A,tau0)$beta_hat
}
