% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knots_sd}
\alias{knots_sd}
\title{function to find the estimated standard deviation for knots
This is a function about how to find the estimated standard deviation for knots in spline linear regression}
\usage{
knots_sd(y, X, Z, tau_out)
}
\arguments{
\item{y}{a vector being the observations of the response variable}

\item{X}{a vec being the variable of interest}

\item{Z}{a matrix being the covariates}

\item{tau_out}{a vector being the estimated knots}
}
\value{
sd_out a vector being the estimated standard deviation for knots
}
\description{
function to find the estimated standard deviation for knots
This is a function about how to find the estimated standard deviation for knots in spline linear regression
}
\examples{
library(lsknots)
set.seed(12345)
n = 1000
eta = 0.5 
sig = 0.03
a=  rnorm(n,0,1)
Z = rnorm(n,0,2) # Z: confounder/other covariate
u = (a+Z)/sqrt(5)
X= pnorm(u,mean=0,sd=1) #X: factor of interest with change-point effects: uniform (0,1)
e = rnorm (n,0,sd=sig)  #e: error term
Y = theta[1] + theta[2]*X + theta[3] * as.numeric(X > theta[7]) * ( X - theta[7] ) + theta[4] * as.numeric(X > theta[8]) * ( X - theta[8] ) + theta[5] * as.numeric(X > theta[9]) * ( X - theta[9] )  + theta[6] * as.numeric(X > theta[10]) * ( X - theta[10] )+ eta*Z+ e 
tau_out = knots_find(Y,A,tau00)
sdout = knots_sd(Y, X, as.matrix(Z),tau_out)
}
