% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_plot.R
\name{ls_plot}
\alias{ls_plot}
\title{Convolution in R
This is a function about how to calculate convolution of two vectors in R.}
\usage{
ls_plot(Y, X, Z, tau00, quantile = 1.96)
}
\arguments{
\item{Y}{a vector being the observations of the response variable}

\item{X}{a vec being the variable of interest}

\item{Z}{a matrix being the covariates}

\item{tau00}{a vector being the estimated knots}

\item{quantile}{the critical value based on your interest, default is 1.96}
}
\value{
a ggplot2 object
}
\description{
Convolution in R
This is a function about how to calculate convolution of two vectors in R.
}
\examples{
library(lsknots)
set.seed(123456)
n=200#sample size
theta = c(5, -1, 3, 5,-9,6 , 0.3, 0.6, 0.8, 0.9)
tau00=c(0.3,0.5,0.85,0.95)
theta = c(9, -5, 6, 7,-8,-3 , 0.1, 0.3, 0.5, 0.7)
tau00=c(0.05,0.45,0.55,0.8)
eta = 0.5 
sig = 0.03
a=  rnorm(n,0,1)
Z = rnorm(n,0,2) 
u = (a+Z)/sqrt(5)
X= pnorm(u,mean=0,sd=1) 
e = rnorm (n,0,sd=sig)  
Y = theta[1] + theta[2]*X + theta[3]*as.numeric(X > theta[7]) * ( X - theta[7] ) + theta[4] * as.numeric(X > theta[8]) * ( X - theta[8] ) + theta[5] * as.numeric(X > theta[9]) * ( X - theta[9] )  + theta[6] * as.numeric(X > theta[10]) * ( X - theta[10] )+ eta*Z+ e 
ls_plot(Y,X,Z,tau00,1.96)
}
